Copy all files from Learn2track

+++ bash_utilities/l2t_organize_from_tractoflow.sh
@@ -3,22 +3,27 @@
 ###############################################################################
 # Your tree should look like:                                                 #
 # derivatives                                                                 #
-#    ├── original (ex, tractoflow input)                                      #
 #    └── tractoflow_output                                                    #
+#    └── ...
 #    └── dwi_ml_ready: will be created now.                                   #
 #                                                                             #
 # This script will create symlinks in dwi_ml_ready, pointing to data from     #
 # tractoflow for each subject:                                                #
 #       └── dwi:                                                              #
 #           └── dwi: Resample/dwi_resample (last dwi output from tractoflow)  #
+#           └── fodf: FODF_Metrics/fodf                                       #
 #           └── bval: Eddy/bval_eddy                                          #
 #           └── bvec: Eddy/bvec_eddy                                          #
 #           └── fa: DTI_metrics/fa                                            #
 #       └── anat :                                                            #
-#           └── t1_tractoflow: Register_T1/t1_warp                            #
+#           └── t1_tractoflow: Register_T1/t1_warped                          #
 #               (It is the last T1 output from tractoflow.)                   #
 #       └── masks :                                                           #
 #           └── wm: Segment_Tissues/mask_wm                                   #
+#       └── bundles  :                                                        #
+#           └── *: Tracking*/*.trk                                            #
+#               or Tracking*/*/*.trk                                          #
+#               (names become Tracking*_*_*.trk)                              #
 #                                                                             #
 # If you need something else for your model, you can modify this script.      #
 #                                                                             #
@@ -84,36 +89,54 @@ do
     mkdir $dwi_ml_folder/$subjid/anat
     mkdir $dwi_ml_folder/$subjid/dwi
     mkdir $dwi_ml_folder/$subjid/masks
-    mkdir $dwi_ml_folder/$subjid/wholebrain
+    mkdir $dwi_ml_folder/$subjid/tractograms
 
     # dwi/dwi:
+    echo "  File: dwi_resampled"
     if [ ! -f $tractoflow_folder/$subjid/Resample_DWI/${subjid}__dwi_resampled.nii.gz ]; then echo "Subject's DWI not found"; exit 1; fi
-    ln -s $tractoflow_folder/$subjid/Resample_DWI/${subjid}__dwi_resampled.nii.gz $dwi_ml_folder/$subjid/dwi/dwi_tractoflow.nii.gz
+    ln -s $tractoflow_folder/$subjid/Resample_DWI/${subjid}__dwi_resampled.nii.gz $dwi_ml_folder/$subjid/dwi/dwi.nii.gz
+
+    # dwi/fodf:
+    echo "  File: fodf"
+    if [ ! -f $tractoflow_folder/$subjid/FODF_Metrics/${subjid}__fodf.nii.gz ]; then echo "Subject's FODF not found"; exit 1; fi
+    ln -s $tractoflow_folder/$subjid/FODF_Metrics/${subjid}__fodf.nii.gz $dwi_ml_folder/$subjid/dwi/fodf.nii.gz
 
     # dwi/bval
+    echo "  File: bval"
     if [ ! -f $tractoflow_folder/$subjid/Eddy/${subjid}__bval_eddy ]; then echo "Subject's bval not found"; exit 1; fi
-    ln -s $tractoflow_folder/$subjid/Eddy/${subjid}__bval_eddy $dwi_ml_folder/$subjid/dwi/bval_tractoflow
+    ln -s $tractoflow_folder/$subjid/Eddy/${subjid}__bval_eddy $dwi_ml_folder/$subjid/dwi/bval
 
     # dwi/bvec
+    echo "  File: bvec"
     if [ ! -f $tractoflow_folder/$subjid/Eddy/${subjid}__dwi_eddy_corrected.bvec ]; then echo "Subject's bvec not found"; exit 1; fi
-    ln -s $tractoflow_folder/$subjid/Eddy/${subjid}__dwi_eddy_corrected.bvec $dwi_ml_folder/$subjid/dwi/bvec_tractoflow
+    ln -s $tractoflow_folder/$subjid/Eddy/${subjid}__dwi_eddy_corrected.bvec $dwi_ml_folder/$subjid/dwi/bvec.bvec
 
     # dwi/FA
+    echo "  File: fa"
     if [ ! -f $tractoflow_folder/$subjid/DTI_Metrics/${subjid}__fa.nii.gz ]; then echo "Subject's FA not found"; exit 1; fi
     ln -s $tractoflow_folder/$subjid/DTI_Metrics/${subjid}__fa.nii.gz $dwi_ml_folder/$subjid/dwi/fa.nii.gz
 
     # anat/T1:
+    echo "  File: t1"
     if [ ! -f $tractoflow_folder/$subjid/Register_T1/${subjid}__t1_warped.nii.gz ]; then echo "Subject's T1 not found"; exit 1; fi
-    ln -s $tractoflow_folder/$subjid/Register_T1/${subjid}__t1_warped.nii.gz $dwi_ml_folder/$subjid/anat/t1_tractoflow.nii.gz
+    ln -s $tractoflow_folder/$subjid/Register_T1/${subjid}__t1_warped.nii.gz $dwi_ml_folder/$subjid/anat/t1.nii.gz
 
     # mask/wm
+    echo "  File: wm"
     if [ ! -f $tractoflow_folder/$subjid/Segment_Tissues/${subjid}__map_wm.nii.gz ]; then echo "Subject's WM map not found"; exit 1; fi
     ln -s $tractoflow_folder/$subjid/Segment_Tissues/${subjid}__map_wm.nii.gz $dwi_ml_folder/$subjid/masks/wm.nii.gz
 
     # tractograms:
-    # Modify this based on your tractogram's name. Or skip and use organaize_from_recobundles.sh.
-    #if [ ! -f $tractoflow_folder/$subjid/*Tracking/${subjid}__*tracking*.trk ]; then echo "Subject's tractogram not found!"; exit 1; fi
-    #ln -s $tractoflow_folder/$subjid/*Tracking/${subjid}__*tracking*.trk $dwi_ml_folder/$subjid/bundles/tractoflow_wholebrain.trk
+    echo "  Tractograms:"
+    tracking_files=$(ls $tractoflow_folder/$subjid/Tracking*/*.trk 2>/dev/null)  # 2>... : to suppress error "No such file or directory"
+    other_files=$(ls $tractoflow_folder/$subjid/Tracking*/*/*.trk 2>/dev/null)
+    for tracking_file in $tracking_files $other_files
+    do
+      echo "    File: $tracking_file"  # File contains the whole path.
+      filename=$(echo ${tracking_file#$tractoflow_folder/$subjid/} | tr / _)
+      echo "       New filename: $filename"
+      ln -s $tracking_file $dwi_ml_folder/$subjid/tractograms/$filename
+    done
 
 done
